check error on instance describe for InstanceNotFound

+++ core/autoscaling.go
@@ -9,6 +9,7 @@ import (
 	"time"
 
 	"github.com/aws/aws-sdk-go/aws"
+	"github.com/aws/aws-sdk-go/aws/awserr"
 	"github.com/aws/aws-sdk-go/service/autoscaling"
 	"github.com/aws/aws-sdk-go/service/ec2"
 )
@@ -598,6 +599,15 @@ func (a *autoScalingGroup) getAnySpotInstance() *instance {
 	return a.getInstance(nil, false, false)
 }
 
+func isInstanceNotFound(err error) bool {
+	if aerr, ok := err.(awserr.Error); ok {
+		if aerr.Code() == "InvalidInstanceID.NotFound" {
+			return true
+		}
+	}
+	return false
+}
+
 func (a *autoScalingGroup) getInstanceState(instanceID *string) int64 {
 	svc := a.region.services.ec2
 	input := &ec2.DescribeInstanceStatusInput{
@@ -609,8 +619,13 @@ func (a *autoScalingGroup) getInstanceState(instanceID *string) int64 {
 
 	out, err := svc.DescribeInstanceStatus(input)
 	if err != nil {
-		logger.Println("Error describing instance status:", *instanceID, err)
-		return 0
+		if isInstanceNotFound(err) {
+			logger.Println("Instance not found:", *instanceID)
+			return 48
+		} else {
+			logger.Println("Error describing instance status:", *instanceID, err)
+			return 0
+		}
 	}
 
 	if len(out.InstanceStatuses) > 0 {
@@ -621,17 +636,27 @@ func (a *autoScalingGroup) getInstanceState(instanceID *string) int64 {
 	return 48
 }
 
-func (a *autoScalingGroup) cancelSIRAndTerminateInstance(instanceID *string, sirRequestID *string) {
+func (a *autoScalingGroup) terminateInstance(instanceID *string, instanceState int64) bool {
 	svc := a.region.services.ec2
-	if instanceID != nil {
+	if instanceID != nil && instanceState > 48 {
 		_, err := svc.TerminateInstances(&ec2.TerminateInstancesInput{
 			InstanceIds: []*string{instanceID},
 		})
 		if err != nil {
-			return
+			return false
 		}
 	}
+	return true
+}
 
+func (a *autoScalingGroup) cancelSIRAndTerminateInstance(instanceID *string, sirRequestID *string, instanceState int64) {
+
+	if !a.terminateInstance(instanceID, instanceState) {
+		logger.Println("Unable to terminate instance:", *instanceID)
+		return
+	}
+
+	svc := a.region.services.ec2
 	_, err := svc.CancelSpotInstanceRequests(
 		&ec2.CancelSpotInstanceRequestsInput{
 			SpotInstanceRequestIds: []*string{sirRequestID},
@@ -669,7 +694,7 @@ func (a *autoScalingGroup) processUnattachedInstance(req *spotInstanceRequest, i
 		return false, false
 	} else if spotInstanceRunning > 16 {
 		// stopped, terminated, shutting-dowm etc,
-		a.cancelSIRAndTerminateInstance(instanceID, req.SpotInstanceRequestId)
+		a.cancelSIRAndTerminateInstance(instanceID, req.SpotInstanceRequestId, spotInstanceRunning)
 		// processNextSIR = true
 		return true, false
 	} else {
